# src/main/resources/com/sonicle/webtop/core/openapi-v1.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "WebTop Core",
    "description": "https://github.com/OpenAPITools/openapi-generator/issues/10010\nhttps://github.com/stoplightio/prism/discussions/523\nhttps://github.com/OpenAPITools/openapi-generator/issues/3172",
    "contact": {
      "email": "dev-team@sonicle.com"
    },
    "version": "v1"
  },
  "servers": [
    {
      "url": "/"
    }
  ],
  "security": [
    {
      "auth-basic": []
    },
    {
      "auth-bearer": []
    }
  ],
  "paths": {
    "/dav/principals/{profileUsername}": {
      "get": {
        "tags": [
          "dav-principals"
        ],
        "summary": "Returns principal info",
        "operationId": "getDavPrincipalInfo",
        "parameters": [
          {
            "name": "profileUsername",
            "in": "path",
            "description": "Full profile username",
            "required": true,
            "schema": {
              "type": "string"
            },
            "example": "john.doe@sonicle.com"
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PrincipalInfo"
                }
              }
            }
          },
          "400": {
            "description": "Invalid parameter",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "404": {
            "description": "Principal not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/principals/{profileUsername}": {
      "get": {
        "tags": [
          "principals"
        ],
        "summary": "Returns principal info",
        "operationId": "getPrincipalInfo",
        "parameters": [
          {
            "name": "profileUsername",
            "in": "path",
            "description": "Full profile username (user@domain.tld)",
            "required": true,
            "schema": {
              "type": "string"
            },
            "example": "john.doe@sonicle.com"
          },
          {
            "name": "permRefs",
            "in": "query",
            "description": "Permissions to evaluate",
            "style": "form",
            "explode": false,
            "schema": {
              "type": "array",
              "items": {
                "type": "string"
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PrincipalInfo"
                }
              }
            }
          },
          "400": {
            "description": "Invalid parameter",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "Invalid credentials",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "404": {
            "description": "Principal not found"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/users": {
      "get": {
        "summary": "List all enabled users",
        "operationId": "getUsers",
        "parameters": [
          {
            "name": "targetProfileId",
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/LegacyUser"
                  }
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        },
        "x-codegen-request-body-name": "body",
        "deprecated": true
      }
    },
...
    "/mail-bridge/relays": {
      "get": {
        "summary": "List MailBridge Relays for a Domain",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Relay"
                  }
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        },
        "operationId": "mailBridgeListRelays",
        "description": "Lists Domain's MailBridge Relays.",
        "parameters": [
          {
            "$ref": "#/components/parameters/domainId"
          }
        ],
        "x-codegen-request-body-name": "body",
        "tags": [
          "mail-bridge-relays"
        ]
      },
      "parameters": []
    },
    "/mail-bridge/relays/{webtop_profile_id}/auth-state": {
      "parameters": [
        {
          "schema": {
            "type": "string"
          },
          "name": "webtop_profile_id",
          "in": "path",
          "required": true
        }
      ],
      "post": {
        "summary": "Update auth state of relay",
        "operationId": "mailBridgeUpdateRelayAuthState",
        "responses": {
          "200": {
            "description": "OK"
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        },
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "in": "query",
            "name": "state",
            "description": "auth state value (UN,OK,ER)"
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "string",
                "description": "Activated string, result of manual (offline) activation."
              }
            }
          },
          "description": ""
        },
        "description": "Updates auth state of relay.",
        "x-codegen-request-body-name": "body",
        "tags": [
          "mail-bridge-relays"
        ]
      }
    },
    "/mail-bridge/fetchers": {
      "get": {
        "summary": "List MailBridge Fetchers for a Domain",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Fetcher"
                  }
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        },
        "operationId": "mailBridgeListFetchers",
        "description": "Lists Domain's MailBridge Fetchers.",
        "parameters": [
          {
            "$ref": "#/components/parameters/domainId"
          }
        ],
        "x-codegen-request-body-name": "body",
        "tags": [
          "mail-bridge-fetchers"
        ]
      },
      "parameters": []
    },
    "/mail-bridge/fetchers/{webtop_profile_id}/auth-state": {
      "parameters": [
        {
          "schema": {
            "type": "string"
          },
          "name": "webtop_profile_id",
          "in": "path",
          "required": true
        }
      ],
      "post": {
        "summary": "Update auth state of fetcher",
        "operationId": "mailBridgeUpdateFetcherAuthState",
        "responses": {
          "200": {
            "description": "OK"
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        },
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "in": "query",
            "name": "state",
            "description": "auth state value (UN,OK,ER)"
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "string",
                "description": "Activated string, result of manual (offline) activation."
              }
            }
          },
          "description": ""
        },
        "description": "Updates auth state of fetcher.",
        "x-codegen-request-body-name": "body",
        "tags": [
          "mail-bridge-fetchers"
        ]
      }
    }
  },
  "components": {
    "schemas": {
      "Error": {
        "required": [
          "code"
        ],
        "type": "object",
        "properties": {
          "code": {
            "maximum": 599,
            "minimum": 400,
            "type": "integer"
          },
          "description": {
            "type": "string",
            "example": "Bad query parameter [$size]: Invalid integer value [abc]"
          }
        },
        "description": "This general detailed error structure is used throughout this API."
      },
      "PrincipalInfo": {
        "required": [
          "displayName",
          "languageTag",
          "profileId",
          "profileUsername",
          "timezoneId"
        ],
        "type": "object",
        "properties": {
          "profileId": {
            "type": "string",
            "description": "Internal profile ID (user@wtdomain)"
          },
          "profileUsername": {
            "type": "string",
            "description": "Full profile username (user@domain.tld)"
          },
          "displayName": {
            "type": "string",
            "description": "Associated display name"
          },
          "emailAddress": {
            "type": "string",
            "description": "Associated email address"
          },
          "timezoneId": {
            "type": "string",
            "description": "Timezone ID",
            "example": "Europe/Rome"
          },
          "languageTag": {
            "type": "string",
            "description": "Language identifier (language_country)",
            "example": "en_EN"
          },
          "evalPermRefs": {
            "type": "array",
            "items": {
              "type": "boolean"
            }
          }
        },
        "title": ""
      },
      "LegacyUser": {
        "required": [
          "displayName",
          "enabled",
          "id",
          "uid"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "uid": {
            "type": "string"
          },
          "enabled": {
            "type": "boolean"
          },
          "displayName": {
            "type": "string"
          }
        }
      },
...
      "Relay": {
        "title": "Relay",
        "type": "object",
        "description": "mailbridge relay",
        "properties": {
          "webtopProfileId": {
            "type": "string"
          },
          "debug": {
            "type": "boolean"
          },
          "host": {
            "type": "string"
          },
          "port": {
            "type": "integer"
          },
          "protocol": {
            "type": "string"
          },
          "username": {
            "type": "string"
          },
          "password": {
            "type": "string"
          },
          "matcher": {
            "type": "string"
          },
          "ssl": {
            "type": "boolean"
          },
          "starttls": {
            "type": "boolean"
          },
          "authState": {
            "type": "string"
          }
        }
      },
      "Fetcher": {
        "title": "Fetcher",
        "type": "object",
        "description": "mailbridge fetcher",
        "properties": {
          "webtopProfileId": {
            "type": "string"
          },
          "deleteOnForward": {
            "type": "boolean"
          },
          "forwardAddress": {
            "type": "string"
          },
          "host": {
            "type": "string"
          },
          "port": {
            "type": "integer"
          },
          "protocol": {
            "type": "string"
          },
          "username": {
            "type": "string"
          },
          "password": {
            "type": "string"
          },
          "authState": {
            "type": "string"
          }
        }
      },
...
    },
    "securitySchemes": {
      "auth-basic": {
        "type": "http",
        "scheme": "basic"
      },
      "auth-bearer": {
        "type": "http",
        "scheme": "bearer"
      },
      "auth-apikey-bearer": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "APIKey",
        "description": "API-Key Authentication: specifies the secret token (bearer and username are both required)"
      },
      "auth-apikey-username": {
        "type": "apiKey",
        "in": "header",
        "name": "X-Auth-Username",
        "description": "API-Key Authentication: specifies target username (bearer and username are both required)"
      }
    },
    "parameters": {
      "domainId": {
        "name": "domain_id",
        "in": "query",
        "required": true,
        "schema": {
          "type": "string"
        },
        "description": "The domain ID"
      },
...
    },
...
  }
}